+++ src/main/java/org/folio/service/ConsortiaService.java
@@ -8,4 +8,5 @@ import org.folio.domain.dto.TenantCollection;
 public interface ConsortiaService {
   TenantCollection getAllConsortiumTenants(String consortiumId);
   Collection<Tenant> getAllConsortiumTenants();
+  String getCentralTenantId();
 }