+++ src/test/java/org/folio/listener/KafkaEventListenerTest.java
@@ -8,6 +8,8 @@ import static org.mockito.Mockito.verify;
 import java.util.Map;
 
 import org.folio.listener.kafka.KafkaEventListener;
+import org.folio.service.ConsortiaService;
+import org.folio.service.impl.LoanEventHandler;
 import org.folio.service.impl.RequestBatchUpdateEventHandler;
 import org.folio.service.impl.RequestEventHandler;
 import org.folio.service.impl.UserEventHandler;
@@ -15,6 +17,7 @@ import org.folio.service.impl.UserGroupEventHandler;
 import org.folio.spring.service.SystemUserScopedExecutionService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.messaging.MessageHeaders;
@@ -24,6 +27,8 @@ class KafkaEventListenerTest {
   @Mock
   RequestEventHandler requestEventHandler;
   @Mock
+  LoanEventHandler loanEventHandler;
+  @Mock
   RequestBatchUpdateEventHandler requestBatchEventHandler;
   @Mock
   SystemUserScopedExecutionService systemUserScopedExecutionService;
@@ -31,14 +36,15 @@ class KafkaEventListenerTest {
   UserGroupEventHandler userGroupEventHandler;
   @Mock
   UserEventHandler userEventHandler;
+  @Mock
+  ConsortiaService consortiaService;
+  @InjectMocks
+  KafkaEventListener kafkaEventListener;
 
   @Test
   void shouldHandleExceptionInEventHandler() {
     doThrow(new NullPointerException("NPE")).when(systemUserScopedExecutionService)
       .executeAsyncSystemUserScoped(any(), any());
-    KafkaEventListener kafkaEventListener = new KafkaEventListener(requestEventHandler,
-      requestBatchEventHandler, systemUserScopedExecutionService, userGroupEventHandler,
-      userEventHandler);
     kafkaEventListener.handleRequestEvent("{}",
       new MessageHeaders(Map.of(TENANT, "default".getBytes())));