+++ src/test/java/org/folio/service/RequestEventHandlerTest.java
@@ -8,7 +8,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.util.Optional;
-import java.util.UUID;
 
 import org.folio.api.BaseIT;
 import org.folio.listener.kafka.KafkaEventListener;
@@ -35,8 +34,7 @@ class RequestEventHandlerTest extends BaseIT {
   void handleRequestUpdateTest() {
     when(ecsTlrRepository.findBySecondaryRequestId(any())).thenReturn(Optional.of(getEcsTlrEntity()));
     doNothing().when(dcbService).createLendingTransaction(any());
-    eventListener.handleRequestEvent(REQUEST_UPDATE_EVENT_SAMPLE, getMessageHeaders(
-      TENANT_ID_CONSORTIUM, UUID.randomUUID().toString()));
+    eventListener.handleRequestEvent(REQUEST_UPDATE_EVENT_SAMPLE, buildKafkaHeaders(TENANT_ID_CONSORTIUM));
     verify(ecsTlrRepository).findBySecondaryRequestId(any());
   }
 }