+++ src/test/java/org/folio/service/UserGroupEventHandlerTest.java
@@ -30,6 +30,7 @@ class UserGroupEventHandlerTest extends BaseEventHandlerTest {
     when(userTenantsService.findFirstUserTenant()).thenReturn(mockUserTenant());
     when(consortiaService.getAllConsortiumTenants(anyString())).thenReturn(mockTenantCollection());
     when(userGroupService.create(any(UserGroup.class))).thenReturn(new UserGroup());
+    when(consortiaService.getCentralTenantId()).thenReturn(CENTRAL_TENANT_ID);
 
     doAnswer(invocation -> {
       ((Runnable) invocation.getArguments()[1]).run();
@@ -38,7 +39,7 @@ class UserGroupEventHandlerTest extends BaseEventHandlerTest {
       any(Runnable.class));
 
     eventListener.handleUserGroupEvent(USER_GROUP_CREATING_EVENT_SAMPLE,
-      getMessageHeaders(TENANT, TENANT_ID));
+      buildKafkaHeaders(CENTRAL_TENANT_ID));
 
     verify(systemUserScopedExecutionService, times(3)).executeAsyncSystemUserScoped(anyString(),
       any(Runnable.class));
@@ -50,6 +51,7 @@ class UserGroupEventHandlerTest extends BaseEventHandlerTest {
     when(userTenantsService.findFirstUserTenant()).thenReturn(mockUserTenant());
     when(consortiaService.getAllConsortiumTenants(anyString())).thenReturn(mockTenantCollection());
     when(userGroupService.update(any(UserGroup.class))).thenReturn(new UserGroup());
+    when(consortiaService.getCentralTenantId()).thenReturn(CENTRAL_TENANT_ID);
 
     doAnswer(invocation -> {
       ((Runnable) invocation.getArguments()[1]).run();
@@ -58,7 +60,7 @@ class UserGroupEventHandlerTest extends BaseEventHandlerTest {
       any(Runnable.class));
 
     eventListener.handleUserGroupEvent(USER_GROUP_UPDATING_EVENT_SAMPLE,
-      getMessageHeaders(TENANT, TENANT_ID));
+      buildKafkaHeaders(CENTRAL_TENANT_ID));
 
     verify(systemUserScopedExecutionService, times(3))
       .executeAsyncSystemUserScoped(anyString(), any(Runnable.class));